Add dry/wet mix control

+++ src/global.h
@@ -39,7 +39,7 @@ namespace VST {
     static const FUID ControllerUID( 0x73A7B7C0, 0x1AD743C1, 0xBFBFD9F4, 0x5F5A04E1 );
 
     static const float DEFAULT_SAMPLE_RATE = 44100.f; // updated in vst.cpp#setupProcessing
-    static const float DEFAULT_BUFFER_SIZE = 2048; // updated in vst.cpp#setupProcessing
+    static const int DEFAULT_BUFFER_SIZE = 2048; // updated in vst.cpp#setupProcessing
 
     static const float PI     = 3.141592653589793f;
     static const float TWO_PI = PI * 2.f;
